speedup installer

+++ utils/common.sh
@@ -193,6 +193,58 @@ function get_os_information() {
     echo -e "[$done_format]"
 }
 
+
+apt_ensure(){
+    __doc__="
+    Checks to see if the packages are installed and installs them if needed.
+
+    The main reason to use this over normal apt install is that it avoids sudo
+    if we already have all requested packages.
+
+    Args:
+        *ARGS : one or more requested packages
+
+    Environment:
+        UPDATE : if this is populated also runs and apt update
+
+    Example:
+        apt_ensure git curl htop
+    "
+    # Note the $@ is not actually an array, but we can convert it to one
+    # https://linuxize.com/post/bash-functions/#passing-arguments-to-bash-functions
+    ARGS=("$@")
+    MISS_PKGS=()
+    HIT_PKGS=()
+    _SUDO=""
+    if [ "$(whoami)" != "root" ]; then
+        # Only use the sudo command if we need it (i.e. we are not root)
+        _SUDO="sudo "
+    fi
+    # shellcheck disable=SC2068
+    for PKG_NAME in ${ARGS[@]}
+    do
+        # Check if the package is already installed or not
+        if dpkg -l "$PKG_NAME" | grep "^ii *$PKG_NAME" > /dev/null; then
+            echo "Already have PKG_NAME='$PKG_NAME'"
+            # shellcheck disable=SC2268,SC2206
+            HIT_PKGS=(${HIT_PKGS[@]} "$PKG_NAME")
+        else
+            echo "Do not have PKG_NAME='$PKG_NAME'"
+            # shellcheck disable=SC2268,SC2206
+            MISS_PKGS=(${MISS_PKGS[@]} "$PKG_NAME")
+        fi
+    done
+    # Install the packages if any are missing
+    if [ "${#MISS_PKGS}" -gt 0 ]; then
+        if [ "${UPDATE}" != "" ]; then
+            $_SUDO apt update -y
+        fi
+        DEBIAN_FRONTEND=noninteractive $_SUDO apt install --no-install-recommends -y "${MISS_PKGS[@]}"
+    else
+        echo "No missing packages"
+    fi
+}
+
 # Install packages required by the installer based on retrieved information
 # from get_os_information() function. If the operating system is not supported then
 # the installer will exit with a message.
@@ -207,8 +259,7 @@ function required_packages() {
 
     case "$DISTRO_NAME" in
     debian | ubuntu | raspbian | linuxmint | zorin)
-        apt-get update &>>"$LOG_FILE"
-        apt-get install --no-install-recommends -y python3 python3-dev python3-pip python3-venv whiptail expect jq "${extra_packages[@]}" &>>"$LOG_FILE"
+        UPDATE=1 apt_ensure python3 python3-dev python3-pip python3-venv whiptail expect jq "${extra_packages[@]}" &>>"$LOG_FILE"
         ;;
     fedora)
         dnf install -y python3 python3-devel python3-pip python3-virtualenv newt expect jq "${extra_packages[@]}" &>>"$LOG_FILE"
@@ -244,16 +295,40 @@ function create_python_venv() {
         fi
     fi
 
+    # two new config variables introduced in this PR.
+    # hard coding for now, pending review.
+    # Defaulting use-uv to true because it is stable and
+    # significantly faster than pip installing dependencies
+    # it may make sense to insist on the use of uv.
+    # We will default REUSED_CACHED_ARTIFACTS to false as
+    # it is mainly useful for debugging.
+
+    USE_UV="true"
+    REUSED_CACHED_ARTIFACTS="false"
+
     if [ -d "$VENV_PATH" ]; then
-        # Make sure everything is clean before starting.
-        rm -rf "$VENV_PATH" /root/.ansible &>>"$LOG_FILE"
+        if [ "$REUSED_CACHED_ARTIFACTS" != "true" ]; then
+            # Question: should we use RUN_AS_HOME/.ansible instead?
+            # Make sure everything is clean before starting.
+            rm -rf "$VENV_PATH" /root/.ansible &>>"$LOG_FILE"
+        fi
     fi
+
     python3 -m venv "$VENV_PATH" &>>"$LOG_FILE"
 
     # shellcheck source=/dev/null
     source "$VENV_PATH/bin/activate"
 
-    pip3 install --upgrade pip setuptools &>>"$LOG_FILE"
+    if [ "$USE_UV" == "true" ]; then
+        export PIP_COMMAND="uv pip"
+        if ! type uv ; then
+            pip3 install "uv>=0.4.10"
+        fi
+    else
+        export PIP_COMMAND="pip3"
+    fi
+
+    $PIP_COMMAND install --upgrade pip setuptools &>>"$LOG_FILE"
     chown "$RUN_AS":"$(id -ng "$RUN_AS")" "$VENV_PATH" "${RUN_AS_HOME}/.venvs" &>>"$LOG_FILE"
     echo -e "[$done_format]"
 }
@@ -265,7 +340,7 @@ function install_ansible() {
     echo -ne "➤ Installing Ansible requirements in Python virtualenv... "
     ANSIBLE_VERSION="9.2.0"
     [ "$(ver "$PYTHON")" -lt "$(ver 3.10)" ] && ANSIBLE_VERSION="8.7.0"
-    pip3 install ansible=="$ANSIBLE_VERSION" docker==7.1.0 requests==2.31.0 &>>"$LOG_FILE"
+    $PIP_COMMAND install ansible=="$ANSIBLE_VERSION" docker==7.1.0 requests==2.31.0 &>>"$LOG_FILE"
     ansible-galaxy collection install -r ansible/requirements.yml &>>"$LOG_FILE"
     echo -e "[$done_format]"
 }
